ECDSA.leftmost_bits: Fixed a major bug in it.
It was failing to accurately detect which cases were unimplemented,
so it was probably failing to raise the exception in cases whre it
should.  It is weird that our tests have not been catching this, so
I need to make a better test for one of smaller curves that forces
leftmost_bits to actually do some bit shifting.